merubah tipe data double menjadi int

+++ app/src/main/java/com/example/ucproomdatabase/data/entity/Barang.kt
@@ -9,7 +9,7 @@ data class Barang(
     val idbarang: String,
     val namabarang: String,
     val deskripsi: String,
-    val harga: Double,
+    val harga: Int,
     val stok: Int,
     val namasuplier: String
 )
